requirements.txt
anyio==4.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.99.1
fastapi-utils==0.2.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
groq==0.4.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
lxml==5.2.0
lxml_html_clean==0.1.0
MarkupSafe==2.1.5
motor==3.4.0
newspaper3k==0.2.8
nltk==3.8.1
openai==1.14.3
orjson==3.10.0
pillow==10.3.0
pydantic==1.10.14
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.52
starlette==0.27.0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.2
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
